publisher/requirements.txt
google-cloud-pubsub
python-dotenv
sqlalchemy
pandas
python-dotenv
psycopg2-binary
PyYAML
